+++ src/mlsquare/__init__.py
@@ -17,4 +17,4 @@ __license__ = "mit"
 
 from .core import dope
 from .base import registry
-from .architectures import sklearn
+from .architectures import sklearn, irt
